+++ react/components/LocalQuery.js
@@ -3,7 +3,7 @@ import { Query } from 'react-apollo'
 import { withRuntimeContext } from 'vtex.render-runtime'
 import { Queries } from 'vtex.store-resources'
 
-import { SORT_OPTIONS } from './OrderBy'
+import { SORT_OPTIONS } from '../OrderBy'
 
 const DEFAULT_PAGE = 1