+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.10.0] - 2019-01-29
+### Added
+- Add `order-by` and `total-products` on `search-result` interface.
+
+### Changed
+- Now, `filter-navigator` block is an allowed block.
+
 ## [3.9.1] - 2019-01-26
 ### Fixed
 - Normalized product now get the sum of skus `AvailableQuantity`.